- Found a solution to the APPINFO initialization Seems I have found a way to get around the issue I faced yesterday and now I know why.
The reason why I couldn't able to copy the a string
into a buffer is that zig have no idea how long the
array really is by looking at a many item pointer
or slices at compile time while requiring strict
definition to the array size.

The solution is to call that function as runtime
instead of stating a predefined function.